+++ src/Checkbox/Checkbox.js
@@ -8,7 +8,6 @@ import Label from '../Label';
 import styles from './Checkbox.scss';
 import WixComponent from '../BaseComponents/WixComponent';
 import {withFocusable, focusableStates} from '../common/Focusable';
-import deprecationLog from '../utils/deprecationLog';
 
 /** a simple WixStyle checkbox */
 class Checkbox extends WixComponent {
@@ -43,20 +42,9 @@ class Checkbox extends WixComponent {
     }
   };
 
-  warnAboutDeprecations = () => {
-    if (this.props.size === 'large') {
-      deprecationLog('Checkbox prop "size" with value "large" is deprecated and will be removed in next major release, please use "medium" size instead');
-    }
-    if (this.props.active !== undefined) {
-      deprecationLog('Checkbox prop "active" is deprecated, use "checked" prop instead');
-    }
-  }
-
   _id = `${Checkbox.displayName}-${uniqueId()}`;
 
   render() {
-    this.warnAboutDeprecations();
-
     const {
       id = this._id,
       checked,
@@ -64,7 +52,6 @@ class Checkbox extends WixComponent {
       disabled,
       hasError,
       hover,
-      active,
       size,
       onChange,
       children
@@ -77,7 +64,6 @@ class Checkbox extends WixComponent {
           styles.unchecked,
       {
         [styles.hover]: hover,
-        [styles.active]: active,
         [styles.disabled]: disabled,
         [styles.hasError]: hasError
       }