+++ src/Heading/Heading.js
@@ -7,7 +7,9 @@ export const APPEARANCES = {
   H1: 'H1',
   H2: 'H2',
   H3: 'H3',
-  H4: 'H4'
+  H4: 'H4',
+  H5: 'H5',
+  H6: 'H6'
 };
 
 const Heading = ({light, appearance, children, ...rest}) => React.createElement(