+++ stories/Table/DefaultToolbar.js
@@ -76,7 +76,6 @@ const BulkActionsToolbar = props => (
     <ItemGroup position="end">
       <Item layout="button">
         <Button
-          withNewIcons
           theme="whiteblueprimary" prefixIcon={<Upload/>}
           onClick={() => window.alert(`Exporting selectedIds=${props.getSelectedIds()}`)}
           >
@@ -85,7 +84,6 @@ const BulkActionsToolbar = props => (
       </Item>
       <Item layout="button">
         <Button
-          withNewIcons
           theme="whiteblueprimary" prefixIcon={<Duplicate/>}
           onClick={() => window.alert(`Duplicating selectedIds=${props.getSelectedIds()}`)}
           >
@@ -94,7 +92,6 @@ const BulkActionsToolbar = props => (
       </Item>
       <Item layout="button">
         <Button
-          withNewIcons
           theme="whiteblueprimary" prefixIcon={<Edit/>}
           onClick={() => window.alert(`Editing selectedIds=${props.getSelectedIds()}`)}
           >