+++ stories/TableToolbar/BulkActionsExample.js
@@ -33,7 +33,6 @@ const BulkActionsToolbar = props => (
     <ItemGroup position="end">
       <Item layout="button">
         <Button
-          withNewIcons
           theme="whiteblueprimary" prefixIcon={<Upload/>}
           onClick={() => window.alert(`Exporting selectedIds=${props.getSelectedIds()}`)}
           >
@@ -42,7 +41,6 @@ const BulkActionsToolbar = props => (
       </Item>
       <Item layout="button">
         <Button
-          withNewIcons
           theme="whiteblueprimary" prefixIcon={<Duplicate/>}
           onClick={() => window.alert(`Duplicating selectedIds=${props.getSelectedIds()}`)}
           >
@@ -51,7 +49,6 @@ const BulkActionsToolbar = props => (
       </Item>
       <Item layout="button">
         <Button
-          withNewIcons
           theme="whiteblueprimary" prefixIcon={<Edit/>}
           onClick={() => window.alert(`Editing selectedIds=${props.getSelectedIds()}`)}
           >